validate add changed toString using String.format tests removed
